requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.9
django-cors-headers==4.1.0
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
Pillow==10.0.0
psycopg2-binary==2.9.7
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
reportlab==4.0.4
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
